requirements.txt
pymupdf>=1.23.0
anthropic>=0.25.0
python-dotenv>=1.0.0
click>=8.0.0
pydantic>=2.0.0